+++ engine/crates/postgres-types/src/transport/tcp.rs
@@ -1,4 +1,6 @@
 mod conversion;
+mod executor;
+mod transaction;
 
 use self::conversion::json_to_string;
 
@@ -6,8 +8,8 @@ use super::Transport;
 use crate::error::Error;
 use async_trait::async_trait;
 use futures::stream::BoxStream;
-use futures::{pin_mut, StreamExt};
 use serde_json::Value;
+use tokio_postgres::Transaction;
 
 pub struct TcpTransport {
     client: tokio_postgres::Client,
@@ -101,46 +103,21 @@ impl TcpTransport {
 
         Ok(this)
     }
+
+    pub async fn transaction(&mut self) -> crate::Result<Transaction<'_>> {
+        Ok(self.client.transaction().await?)
+    }
 }
 
 #[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
 #[cfg_attr(not(target_arch = "wasm32"), async_trait)]
 impl Transport for TcpTransport {
     fn parameterized_query<'a>(&'a self, query: &'a str, params: Vec<Value>) -> BoxStream<'a, Result<Value, Error>> {
-        Box::pin(async_stream::try_stream! {
-            let params = json_to_string(params);
-            let row_stream = self.client.query_raw_txt(query, params).await?;
-
-            pin_mut!(row_stream);
-
-            while let Some(row) = row_stream.next().await {
-                yield serde_json::from_value(conversion::row_to_json(&row?)?)?;
-            }
-        })
+        executor::query(&self.client, query, params)
     }
 
     async fn parameterized_execute(&self, query: &str, params: Vec<Value>) -> crate::Result<i64> {
-        let params = json_to_string(params);
-        let row_stream = self.client.query_raw_txt(query, params).await?;
-
-        pin_mut!(row_stream);
-
-        while (row_stream.next().await).is_some() {}
-
-        let command_tag = row_stream.command_tag().unwrap_or_default();
-        let mut command_tag_split = command_tag.split(' ');
-        let command_tag_name = command_tag_split.next().unwrap_or_default();
-
-        let row_count = if command_tag_name == "INSERT" {
-            // INSERT returns OID first and then number of rows
-            command_tag_split.nth(1)
-        } else {
-            // other commands return number of rows (if any)
-            command_tag_split.next()
-        }
-        .and_then(|s| s.parse::<i64>().ok());
-
-        Ok(row_count.unwrap_or_default())
+        executor::execute(&self.client, query, params).await
     }
 
     fn connection_string(&self) -> &str {
